requirements.txt
 Core Django
Django==6.0.3
asgiref==3.11.1
sqlparse==0.5.5

# Images
pillow==12.1.1

# Configuration
python-dotenv==1.2.2

# Interface Admin
django-unfold==0.87.0

# API (Flutterwave)
requests==2.33.0
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
urllib3==2.6.3

# API Externe
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.12.1
